requirements.txt
aniso8601==9.0.1
Babel==2.12.1
Beaker==1.12.1
black==23.3.0
BTrees==5.0
certifi==2022.12.7
cffi==1.15.1
Chameleon==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colander==2.0
cornice==6.0.1
cornice-swagger==1.0.1
cssselect==1.2.0
docutils==0.19
fanstatic==1.3
graphene==3.2.1
greenlet==2.0.2
httplib2==0.21.0
hupper==1.11
hypatia==0.4
idna==3.4
iso8601==1.1.0
joblib==1.2.0
lxml==4.9.2
Markdown==3.4.1
multipart==0.2.4
mypy-extensions==1.0.0
nltk==3.8.1
packaging==23.1
PasteDeploy==3.0.1
pathspec==0.11.1
persistent==5.0
Pillow==9.4.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==3.2.0
psycopg-binary==3.1.8
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gettext==4.1
python-magic==0.4.27
pytz==2022.7.1
regex==2022.10.31
repoze.sendmail==4.4.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
tomli==2.0.1
tqdm==4.64.1
transaction==3.0.1
translationstring==1.4
typing_extensions==4.5.0
urllib3==1.26.14
venusian==3.0.0
vk-api==11.9.9
